server/test_jwt.py
#!/usr/bin/env python3
"""
Test JWT Authentication Implementation
"""

import requests
import json

BASE_URL = "http://localhost:8000"

def test_jwt_login():
    """Test JWT login endpoint"""
    print("🔐 Testing JWT Login...")
    
    # Test login with form data (as expected by FastAPI Form)
    login_data = {
        'username': 'admin1',
        'password': 'password123'
    }
    
    try:
        response = requests.post(f"{BASE_URL}/auth/login", data=login_data)
        
        if response.status_code == 200:
            result = response.json()
            print("✅ JWT Login successful!")
            print(f"   Token Type: {result['token_type']}")
            print(f"   User: {result['user']['username']}")
            print(f"   Role: {result['user']['role']}")
            print(f"   Token: {result['access_token'][:50]}...")
            return result['access_token']
        else:
            print(f"❌ JWT Login failed: {response.status_code}")
            print(f"   Response: {response.text}")
            return None
            
    except Exception as e:
        print(f"❌ JWT Login error: {e}")
        return None

def test_jwt_protected_endpoint(token):
    """Test accessing protected endpoint with JWT token"""
    print("\n🔒 Testing JWT Protected Endpoint...")
    
    headers = {
        "Authorization": f"Bearer {token}"
    }
    
    try:
        response = requests.get(f"{BASE_URL}/admin/dashboard", headers=headers)
        
        if response.status_code == 200:
            result = response.json()
            print("✅ JWT Protected endpoint access successful!")
            print(f"   Total Applications: {result['system_stats']['total_applications']}")
            print(f"   Total Users: {result['system_stats']['total_users']}")
            return True
        else:
            print(f"❌ JWT Protected endpoint failed: {response.status_code}")
            print(f"   Response: {response.text}")
            return False
            
    except Exception as e:
        print(f"❌ JWT Protected endpoint error: {e}")
        return False

def test_invalid_token():
    """Test with invalid token"""
    print("\n🚫 Testing Invalid Token...")
    
    headers = {
        "Authorization": "Bearer invalid_token_here"
    }
    
    try:
        response = requests.get(f"{BASE_URL}/admin/dashboard", headers=headers)
        
        if response.status_code == 401:
            print("✅ Invalid token correctly rejected!")
            return True
        else:
            print(f"❌ Invalid token test failed: {response.status_code}")
            print(f"   Response: {response.text}")
            return False
            
    except Exception as e:
        print(f"❌ Invalid token test error: {e}")
        return False

def main():
    """Run JWT authentication tests"""
    print("🧪 JWT Authentication Test Suite")
    print("=" * 50)
    
    # Test 1: Login and get JWT token
    token = test_jwt_login()
    if not token:
        print("\n❌ Cannot proceed without valid token")
        return
    
    # Test 2: Access protected endpoint with valid token
    test_jwt_protected_endpoint(token)
    
    # Test 3: Test invalid token rejection
    test_invalid_token()
    
    print("\n" + "=" * 50)
    print("🎉 JWT Authentication tests completed!")
    print("\nJWT Implementation Status:")
    print("✅ POST /auth/login - Returns JWT tokens")
    print("✅ Bearer token authentication working")
    print("✅ Protected endpoints secured")
    print("✅ Invalid tokens properly rejected")
    print("\n🚀 Your application now uses JWT authentication!")

if __name__ == "__main__":
    main()
