server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pymongo==4.6.0
bcrypt==4.1.2
pydantic==2.5.0
python-multipart==0.0.6
certifi
# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
\# Optional LLM support
groq==0.9.0